+++ hashtagmap-batch/src/main/java/com/songpapeople/hashtagmap/instagram/config/InstagramCrawlingJobConfiguration.java
@@ -3,7 +3,7 @@ package com.songpapeople.hashtagmap.instagram.config;
 import com.songpapeople.hashtagmap.instagram.processor.InstagramBatchProcessor;
 import com.songpapeople.hashtagmap.instagram.writer.InstagramBatchWriter;
 import com.songpapeople.hashtagmap.place.domain.model.Place;
-import com.songpapeople.hashtagmap.scheduler.CrawlingResult;
+import com.songpapeople.hashtagmap.service.CrawlingResult;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
