+++ hashtagmap-batch/src/test/java/com/songpapeople/hashtagmap/instagram/InstagramBatchIntegrationTest.java
@@ -6,8 +6,8 @@ import com.songpapeople.hashtagmap.dto.PostDtos;
 import com.songpapeople.hashtagmap.instagram.domain.repository.InstagramRepository;
 import com.songpapeople.hashtagmap.place.domain.model.Place;
 import com.songpapeople.hashtagmap.place.domain.repository.PlaceRepository;
-import com.songpapeople.hashtagmap.scheduler.CrawlingResult;
-import com.songpapeople.hashtagmap.scheduler.InstagramScheduler;
+import com.songpapeople.hashtagmap.service.CrawlingResult;
+import com.songpapeople.hashtagmap.service.InstagramScheduler;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
