+++ hashtagmap-instagram-service/src/main/java/com/songpapeople/hashtagmap/service/ProxySetter.java
@@ -1,4 +1,4 @@
-package com.songpapeople.hashtagmap.scheduler;
+package com.songpapeople.hashtagmap.service;
 
 import com.songpapeople.hashtagmap.proxy.Proxies;
