+++ hashtagmap-instagram-service/src/test/java/com/songpapeople/hashtagmap/service/CrawlingResultTest.java
@@ -1,4 +1,4 @@
-package com.songpapeople.hashtagmap.scheduler;
+package com.songpapeople.hashtagmap.service;
 
 import com.songpapeople.hashtagmap.MockDataFactory;
 import com.songpapeople.hashtagmap.dto.CrawlingDto;
